made database use event source datastore using the hash_map/btree/ordered_array algs to store the projection

+++ algorithm_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 )
 
-func getInputValues() map[string]int64 {
-	testCases := map[string]int64{
+func getInputValues() map[string]int {
+	testCases := map[string]int{
 		"key1":    1,
 		"key2":    2,
 		"key3":    3,
